requirements.txt
flask
faiss-cpu
sentence-transformers
pypdf
google-generativeai
